refactor: Refactorizar la estructura del código para mejorar la legibilidad y el mantenimiento (#16)
* refactor: Refactorizar la estructura del código para mejorar la legibilidad y el mantenimiento

* fix: Corregir rutas de importación para los módulos de configuración y middleware
