+++ src/helpers/deleteTeacherByUser.js
