+++ src/helpers/teachersEliminatedStatus.js
