+++ src/router/catInventarioRouter.js
